fx-discord-news/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fx-discord-news"
version = "1.0.0"
description = "FX news collection, analysis, and Discord delivery automation system"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "FX News Bot", email = "fx@example.com"}
]
dependencies = [
    "httpx>=0.25.0",
    "feedparser>=6.0.10",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "langdetect>=1.0.9",
    "rapidfuzz>=3.5.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.1",
    "loguru>=0.7.2",
    "tenacity>=8.2.3",
    "click>=8.1.7",
    "anthropic>=0.18.0",
    "openai>=1.10.0",
    "pytz>=2023.3",
    "lxml>=4.9.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-pyyaml>=6.0.0",
    "types-pytz>=2023.3.0",
]

[project.scripts]
fx-news = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "PT", "SIM", "RUF"]
ignore = ["E501", "B008", "RUF003"]
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true